ENH: New accessor for Series `invert_or_not` (#801)
* New accessor

* index method

* simplify a bit

* BOT: auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
