+++ dtoolkit/accessor/series/__init__.py
@@ -10,6 +10,7 @@ from dtoolkit.accessor.series.expand import expand  # noqa: F401
 from dtoolkit.accessor.series.filter_in import filter_in  # noqa: F401
 from dtoolkit.accessor.series.getattr import getattr  # noqa: F401
 from dtoolkit.accessor.series.groupby_index import groupby_index  # noqa: F401
+from dtoolkit.accessor.series.invert_or_not import invert_or_not  # noqa: F401
 from dtoolkit.accessor.series.jenks_bin import jenks_bin  # noqa: F401
 from dtoolkit.accessor.series.jenks_breaks import jenks_breaks  # noqa: F401
 from dtoolkit.accessor.series.len import len  # noqa: F401
